Refactor: Organize use cases into feature-specific packages
This commit refactors the `domain/use-case` module by organizing all use case classes into feature-specific sub-packages. This improves code organization, modularity, and discoverability.

The new package structure is as follows:
- `domain.usecase.grid`
- `domain.usecase.iconpack`
- `domain.usecase.launcherapps`
- `domain.usecase.page`
- `domain.usecase.pin`

All import statements in dependent modules (such as `feature/home`, `feature/edit-grid-item`, and `service`) have been updated to reflect these changes.

Additionally, this commit includes a minor visual enhancement to the `EditPageScreen`, where a background with rounded corners is added to the draggable page previews for better visual separation.
