+++ domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/iconpack/UpdateIconPackInfoByPackageNameUseCase.kt
@@ -15,7 +15,7 @@
  *   limitations under the License.
  *
  */
-package com.eblan.launcher.domain.usecase
+package com.eblan.launcher.domain.usecase.iconpack
 
 import com.eblan.launcher.domain.common.dispatcher.Dispatcher
 import com.eblan.launcher.domain.common.dispatcher.EblanDispatchers
