+++ domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/launcherapps/ChangePackageUseCase.kt
@@ -15,24 +15,21 @@
  *   limitations under the License.
  *
  */
-package com.eblan.launcher.domain.usecase
+package com.eblan.launcher.domain.usecase.launcherapps
 
 import com.eblan.launcher.domain.common.dispatcher.Dispatcher
 import com.eblan.launcher.domain.common.dispatcher.EblanDispatchers
 import com.eblan.launcher.domain.framework.AppWidgetManagerWrapper
 import com.eblan.launcher.domain.framework.FileManager
 import com.eblan.launcher.domain.framework.LauncherAppsWrapper
-import com.eblan.launcher.domain.framework.NotificationManagerWrapper
 import com.eblan.launcher.domain.framework.PackageManagerWrapper
 import com.eblan.launcher.domain.model.AppWidgetManagerAppWidgetProviderInfo
-import com.eblan.launcher.domain.model.ApplicationInfoGridItem
 import com.eblan.launcher.domain.model.EblanAppWidgetProviderInfo
 import com.eblan.launcher.domain.model.EblanShortcutInfo
-import com.eblan.launcher.domain.model.LauncherAppsActivityInfo
-import com.eblan.launcher.domain.model.LauncherAppsShortcutInfo
+import com.eblan.launcher.domain.model.ShortcutConfigGridItem
 import com.eblan.launcher.domain.model.ShortcutInfoGridItem
-import com.eblan.launcher.domain.model.SyncEblanApplicationInfo
 import com.eblan.launcher.domain.model.UpdateApplicationInfoGridItem
+import com.eblan.launcher.domain.model.UpdateShortcutConfigGridItem
 import com.eblan.launcher.domain.model.UpdateShortcutInfoGridItem
 import com.eblan.launcher.domain.model.UpdateWidgetGridItem
 import com.eblan.launcher.domain.model.WidgetGridItem
@@ -40,103 +37,79 @@ import com.eblan.launcher.domain.repository.ApplicationInfoGridItemRepository
 import com.eblan.launcher.domain.repository.EblanAppWidgetProviderInfoRepository
 import com.eblan.launcher.domain.repository.EblanApplicationInfoRepository
 import com.eblan.launcher.domain.repository.EblanShortcutInfoRepository
+import com.eblan.launcher.domain.repository.ShortcutConfigGridItemRepository
 import com.eblan.launcher.domain.repository.ShortcutInfoGridItemRepository
 import com.eblan.launcher.domain.repository.UserDataRepository
 import com.eblan.launcher.domain.repository.WidgetGridItemRepository
+import com.eblan.launcher.domain.usecase.iconpack.UpdateIconPackInfoByPackageNameUseCase
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.currentCoroutineContext
 import kotlinx.coroutines.ensureActive
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.joinAll
-import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import java.io.File
 import javax.inject.Inject
 
-class SyncDataUseCase @Inject constructor(
+class ChangePackageUseCase @Inject constructor(
     private val userDataRepository: UserDataRepository,
-    private val notificationManagerWrapper: NotificationManagerWrapper,
-    private val eblanApplicationInfoRepository: EblanApplicationInfoRepository,
-    private val launcherAppsWrapper: LauncherAppsWrapper,
+    private val packageManagerWrapper: PackageManagerWrapper,
     private val fileManager: FileManager,
-    private val updateIconPackInfosUseCase: UpdateIconPackInfosUseCase,
+    private val eblanApplicationInfoRepository: EblanApplicationInfoRepository,
+    private val updateIconPackInfoByPackageNameUseCase: UpdateIconPackInfoByPackageNameUseCase,
+    private val applicationInfoGridItemRepository: ApplicationInfoGridItemRepository,
+    private val launcherAppsWrapper: LauncherAppsWrapper,
     private val eblanAppWidgetProviderInfoRepository: EblanAppWidgetProviderInfoRepository,
     private val appWidgetManagerWrapper: AppWidgetManagerWrapper,
-    private val packageManagerWrapper: PackageManagerWrapper,
     private val eblanShortcutInfoRepository: EblanShortcutInfoRepository,
-    private val applicationInfoGridItemRepository: ApplicationInfoGridItemRepository,
     private val widgetGridItemRepository: WidgetGridItemRepository,
     private val shortcutInfoGridItemRepository: ShortcutInfoGridItemRepository,
     private val updateEblanShortcutConfigsUseCase: UpdateEblanShortcutConfigsUseCase,
+    private val shortcutConfigGridItemRepository: ShortcutConfigGridItemRepository,
     @param:Dispatcher(EblanDispatchers.IO) private val ioDispatcher: CoroutineDispatcher,
 ) {
-    suspend operator fun invoke() {
+    suspend operator fun invoke(
+        serialNumber: Long,
+        packageName: String,
+    ) {
         withContext(ioDispatcher) {
-            try {
-                notificationManagerWrapper.notifySyncData()
+            if (!userDataRepository.userData.first().experimentalSettings.syncData) return@withContext
 
-                joinAll(
-                    launch {
-                        val launcherAppsActivityInfos = launcherAppsWrapper.getActivityList()
-
-                        updateEblanApplicationInfos(launcherAppsActivityInfos = launcherAppsActivityInfos)
-
-                        updateApplicationInfoGridItems(launcherAppsActivityInfos = launcherAppsActivityInfos)
-                    },
-                    launch {
-                        val appWidgetManagerAppWidgetProviderInfos =
-                            appWidgetManagerWrapper.getInstalledProviders()
-
-                        updateEblanAppWidgetProviderInfos(appWidgetManagerAppWidgetProviderInfos = appWidgetManagerAppWidgetProviderInfos)
-
-                        updateWidgetGridItems(appWidgetManagerAppWidgetProviderInfos = appWidgetManagerAppWidgetProviderInfos)
-                    },
-                    launch {
-                        updateEblanShortcutInfos(launcherAppsShortcutInfos = launcherAppsWrapper.getShortcuts())
-                    },
-                    launch {
-                        updateShortcutInfoGridItems(pinnedLauncherAppsShortcutInfos = launcherAppsWrapper.getPinnedShortcuts())
-                    },
-                )
-            } finally {
-                notificationManagerWrapper.cancelSyncData()
-            }
-        }
-    }
-
-    private suspend fun updateEblanApplicationInfos(launcherAppsActivityInfos: List<LauncherAppsActivityInfo>) {
-        val iconPackInfoPackageName =
-            userDataRepository.userData.first().generalSettings.iconPackInfoPackageName
-
-        val oldSyncEblanApplicationInfos =
-            eblanApplicationInfoRepository.eblanApplicationInfos.first()
-                .map { eblanApplicationInfo ->
-                    SyncEblanApplicationInfo(
-                        serialNumber = eblanApplicationInfo.serialNumber,
-                        componentName = eblanApplicationInfo.componentName,
-                        packageName = eblanApplicationInfo.packageName,
-                        icon = eblanApplicationInfo.icon,
-                        label = eblanApplicationInfo.label,
-                    )
-                }
-
-        val newSyncEblanApplicationInfos =
-            launcherAppsActivityInfos.map { launcherAppsActivityInfo ->
-                currentCoroutineContext().ensureActive()
-
-                val launcherAppsActivityInfoIcon = launcherAppsActivityInfo.activityIcon
-                    ?: launcherAppsActivityInfo.applicationIcon
-
-                val icon = launcherAppsActivityInfoIcon?.let { byteArray ->
+            launcherAppsWrapper.getActivityList(
+                serialNumber = serialNumber,
+                packageName = packageName,
+            ).forEach { launcherAppsActivityInfo ->
+                val icon = launcherAppsActivityInfo.applicationIcon?.let { currentIconByteArray ->
                     fileManager.updateAndGetFilePath(
                         directory = fileManager.getFilesDirectory(FileManager.ICONS_DIR),
                         name = launcherAppsActivityInfo.componentName.replace(
                             "/",
                             "-",
                         ),
-                        byteArray = byteArray,
+                        byteArray = currentIconByteArray,
                     )
                 }
+
+                updateEblanApplicationInfoByPackageName(
+                    componentName = launcherAppsActivityInfo.componentName,
+                    packageName = launcherAppsActivityInfo.packageName,
+                    serialNumber = launcherAppsActivityInfo.serialNumber,
+                    icon = icon,
+                    label = launcherAppsActivityInfo.label,
+                )
+
+                updateEblanAppWidgetProviderInfosByPackageName(
+                    serialNumber = launcherAppsActivityInfo.serialNumber,
+                    packageName = launcherAppsActivityInfo.packageName,
+                    icon = icon,
+                    label = launcherAppsActivityInfo.label,
+                )
+
+                updateShortcutConfigGridItemsByPackageName(
+                    serialNumber = launcherAppsActivityInfo.serialNumber,
+                    packageName = launcherAppsActivityInfo.packageName,
+                    applicationIcon = icon,
+                    applicationLabel = launcherAppsActivityInfo.label,
+                )
 
                 updateEblanShortcutConfigsUseCase(
                     serialNumber = launcherAppsActivityInfo.serialNumber,
@@ -145,99 +118,87 @@ class SyncDataUseCase @Inject constructor(
                     label = launcherAppsActivityInfo.label,
                 )
 
-                SyncEblanApplicationInfo(
-                    serialNumber = launcherAppsActivityInfo.serialNumber,
-                    componentName = launcherAppsActivityInfo.componentName,
+                updateIconPackInfoByPackageNameUseCase(
                     packageName = launcherAppsActivityInfo.packageName,
-                    icon = icon,
-                    label = launcherAppsActivityInfo.label,
+                    componentName = launcherAppsActivityInfo.componentName,
                 )
             }
 
-        if (oldSyncEblanApplicationInfos != newSyncEblanApplicationInfos) {
-            val upsertEblanApplicationInfosToDelete =
-                oldSyncEblanApplicationInfos - newSyncEblanApplicationInfos.toSet()
+            updateEblanShortcutInfosByPackageName(
+                serialNumber = serialNumber,
+                packageName = packageName,
+            )
 
-            eblanApplicationInfoRepository.upsertSyncEblanApplicationInfos(syncEblanApplicationInfos = newSyncEblanApplicationInfos)
-
-            eblanApplicationInfoRepository.deleteSyncEblanApplicationInfos(syncEblanApplicationInfos = upsertEblanApplicationInfosToDelete)
-
-            upsertEblanApplicationInfosToDelete.forEach { eblanApplicationInfoToDelete ->
-                currentCoroutineContext().ensureActive()
-
-                val isUnique = newSyncEblanApplicationInfos.none { newEblanApplicationInfo ->
-                    newEblanApplicationInfo.packageName == eblanApplicationInfoToDelete.packageName &&
-                        newEblanApplicationInfo.serialNumber != eblanApplicationInfoToDelete.serialNumber
-                }
-
-                if (isUnique) {
-                    eblanApplicationInfoToDelete.icon?.let { icon ->
-                        val iconFile = File(icon)
-
-                        if (iconFile.exists()) {
-                            iconFile.delete()
-                        }
-                    }
-
-                    val iconPacksDirectory = File(
-                        fileManager.getFilesDirectory(FileManager.ICON_PACKS_DIR),
-                        iconPackInfoPackageName,
-                    )
-
-                    val iconPackFile =
-                        File(
-                            iconPacksDirectory,
-                            eblanApplicationInfoToDelete.packageName,
-                        )
-
-                    if (iconPackFile.exists()) {
-                        iconPackFile.delete()
-                    }
-                }
-            }
-
-            updateIconPackInfosUseCase(iconPackInfoPackageName = iconPackInfoPackageName)
+            updateShortcutInfoGridItemsByPackageName(
+                serialNumber = serialNumber,
+                packageName = packageName,
+            )
         }
     }
 
-    private suspend fun updateEblanAppWidgetProviderInfos(appWidgetManagerAppWidgetProviderInfos: List<AppWidgetManagerAppWidgetProviderInfo>) {
+    private suspend fun updateEblanApplicationInfoByPackageName(
+        componentName: String,
+        packageName: String,
+        serialNumber: Long,
+        icon: String?,
+        label: String,
+    ) {
+        val eblanApplicationInfo = eblanApplicationInfoRepository.getEblanApplicationInfo(
+            serialNumber = serialNumber,
+            packageName = packageName,
+        )
+
+        if (eblanApplicationInfo != null) {
+            eblanApplicationInfoRepository.updateEblanApplicationInfo(
+                eblanApplicationInfo = eblanApplicationInfo.copy(
+                    componentName = componentName,
+                    icon = icon,
+                    label = label,
+                ),
+            )
+
+            updateApplicationInfoGridItemsByPackageName(
+                serialNumber = serialNumber,
+                packageName = packageName,
+                componentName = componentName,
+                label = label,
+            )
+        }
+    }
+
+    private suspend fun updateEblanAppWidgetProviderInfosByPackageName(
+        serialNumber: Long,
+        packageName: String,
+        icon: String?,
+        label: String,
+    ) {
         if (!packageManagerWrapper.hasSystemFeatureAppWidgets) return
 
+        val appWidgetManagerAppWidgetProviderInfos = appWidgetManagerWrapper.getInstalledProviders()
+
         val oldEblanAppWidgetProviderInfos =
-            eblanAppWidgetProviderInfoRepository.eblanAppWidgetProviderInfos.first()
+            eblanAppWidgetProviderInfoRepository.getEblanAppWidgetProviderInfosByPackageName(
+                packageName = packageName,
+            )
 
         val newEblanAppWidgetProviderInfos =
-            appWidgetManagerAppWidgetProviderInfos.map { appWidgetManagerAppWidgetProviderInfo ->
-                currentCoroutineContext().ensureActive()
-
-                val preview =
-                    appWidgetManagerAppWidgetProviderInfo.preview?.let { byteArray ->
-                        fileManager.updateAndGetFilePath(
-                            directory = fileManager.getFilesDirectory(FileManager.WIDGETS_DIR),
-                            name = appWidgetManagerAppWidgetProviderInfo.componentName.replace(
-                                "/",
-                                "-",
-                            ),
-                            byteArray = byteArray,
-                        )
-                    }
-
-                val label =
-                    packageManagerWrapper.getApplicationLabel(packageName = appWidgetManagerAppWidgetProviderInfo.packageName)
-
-                val icon =
-                    packageManagerWrapper.getApplicationIcon(packageName = appWidgetManagerAppWidgetProviderInfo.packageName)
-                        ?.let { byteArray ->
-                            fileManager.updateAndGetFilePath(
-                                directory = fileManager.getFilesDirectory(FileManager.ICONS_DIR),
-                                name = appWidgetManagerAppWidgetProviderInfo.packageName,
-                                byteArray = byteArray,
-                            )
-                        }
+            appWidgetManagerAppWidgetProviderInfos.filter { appWidgetManagerAppWidgetProviderInfo ->
+                appWidgetManagerAppWidgetProviderInfo.packageName == packageName
+            }.map { appWidgetManagerAppWidgetProviderInfo ->
+                val preview = appWidgetManagerAppWidgetProviderInfo.preview?.let { byteArray ->
+                    fileManager.updateAndGetFilePath(
+                        directory = fileManager.getFilesDirectory(FileManager.WIDGETS_DIR),
+                        name = appWidgetManagerAppWidgetProviderInfo.componentName.replace(
+                            "/",
+                            "-",
+                        ),
+                        byteArray = byteArray,
+                    )
+                }
 
                 EblanAppWidgetProviderInfo(
                     componentName = appWidgetManagerAppWidgetProviderInfo.componentName,
-                    serialNumber = appWidgetManagerAppWidgetProviderInfo.serialNumber,
+                    serialNumber = serialNumber,
                     configure = appWidgetManagerAppWidgetProviderInfo.configure,
                     packageName = appWidgetManagerAppWidgetProviderInfo.packageName,
                     targetCellWidth = appWidgetManagerAppWidgetProviderInfo.targetCellWidth,
@@ -251,7 +212,7 @@ class SyncDataUseCase @Inject constructor(
                     maxResizeHeight = appWidgetManagerAppWidgetProviderInfo.maxResizeHeight,
                     preview = preview,
                     icon = icon,
-                    label = label.toString(),
+                    label = label,
                 )
             }
 
@@ -268,8 +229,6 @@ class SyncDataUseCase @Inject constructor(
             )
 
             eblanAppWidgetProviderInfosToDelete.forEach { eblanAppWidgetProviderInfoToDelete ->
-                currentCoroutineContext().ensureActive()
-
                 eblanAppWidgetProviderInfoToDelete.icon?.let { icon ->
                     val iconFile = File(icon)
 
@@ -287,14 +246,26 @@ class SyncDataUseCase @Inject constructor(
                 }
             }
         }
+
+        updateWidgetGridItemsByPackageName(
+            appWidgetManagerAppWidgetProviderInfos = appWidgetManagerAppWidgetProviderInfos,
+            serialNumber = serialNumber,
+            packageName = packageName,
+        )
     }
 
-    private suspend fun updateEblanShortcutInfos(launcherAppsShortcutInfos: List<LauncherAppsShortcutInfo>?) {
+    private suspend fun updateEblanShortcutInfosByPackageName(
+        serialNumber: Long,
+        packageName: String,
+    ) {
         if (!launcherAppsWrapper.hasShortcutHostPermission) return
 
         val oldEblanShortcutInfos = eblanShortcutInfoRepository.eblanShortcutInfos.first()
 
-        val newEblanShortcutInfos = launcherAppsShortcutInfos?.map { launcherAppsShortcutInfo ->
+        val newEblanShortcutInfos = launcherAppsWrapper.getShortcutsByPackageName(
+            serialNumber = serialNumber,
+            packageName = packageName,
+        )?.map { launcherAppsShortcutInfo ->
             currentCoroutineContext().ensureActive()
 
             val icon = launcherAppsShortcutInfo.icon?.let { byteArray ->
@@ -318,8 +289,7 @@ class SyncDataUseCase @Inject constructor(
         }
 
         if (newEblanShortcutInfos != null && oldEblanShortcutInfos != newEblanShortcutInfos) {
-            val eblanShortcutInfosToDelete =
-                oldEblanShortcutInfos - newEblanShortcutInfos.toSet()
+            val eblanShortcutInfosToDelete = oldEblanShortcutInfos - newEblanShortcutInfos.toSet()
 
             eblanShortcutInfoRepository.upsertEblanShortcutInfos(
                 eblanShortcutInfos = newEblanShortcutInfos,
@@ -329,83 +299,82 @@ class SyncDataUseCase @Inject constructor(
                 eblanShortcutInfos = eblanShortcutInfosToDelete,
             )
 
-            eblanShortcutInfosToDelete.forEach { eblanShortcutInfoToDelete ->
+            eblanShortcutInfosToDelete.forEach { eblanShortcutInfo ->
                 currentCoroutineContext().ensureActive()
 
-                eblanShortcutInfoToDelete.icon?.let { icon ->
-                    val iconFile = File(icon)
+                val isUnique = eblanShortcutInfoRepository.eblanShortcutInfos.first()
+                    .none { newEblanShortcutInfo ->
+                        currentCoroutineContext().ensureActive()
 
-                    if (iconFile.exists()) {
-                        iconFile.delete()
+                        newEblanShortcutInfo.packageName == packageName && newEblanShortcutInfo.serialNumber != serialNumber
+                    }
+
+                if (isUnique) {
+                    eblanShortcutInfo.icon?.let { icon ->
+                        val iconFile = File(icon)
+
+                        if (iconFile.exists()) {
+                            iconFile.delete()
+                        }
                     }
                 }
             }
         }
     }
 
-    private suspend fun updateApplicationInfoGridItems(launcherAppsActivityInfos: List<LauncherAppsActivityInfo>) {
-        val updateApplicationInfoGridItems = mutableListOf<UpdateApplicationInfoGridItem>()
-
-        val deleteApplicationInfoGridItems = mutableListOf<ApplicationInfoGridItem>()
-
-        val applicationInfoGridItems =
-            applicationInfoGridItemRepository.applicationInfoGridItems.first()
-
-        applicationInfoGridItems.filterNot { applicationInfoGridItem ->
-            applicationInfoGridItem.override
-        }.forEach { applicationInfoGridItem ->
-            currentCoroutineContext().ensureActive()
-
-            val launcherAppsActivityInfo =
-                launcherAppsActivityInfos.find { launcherAppsActivityInfo ->
-                    currentCoroutineContext().ensureActive()
-
-                    launcherAppsActivityInfo.packageName == applicationInfoGridItem.packageName &&
-                        launcherAppsActivityInfo.serialNumber == applicationInfoGridItem.serialNumber
-                }
-
-            if (launcherAppsActivityInfo != null) {
-                updateApplicationInfoGridItems.add(
+    private suspend fun updateApplicationInfoGridItemsByPackageName(
+        serialNumber: Long,
+        packageName: String,
+        componentName: String?,
+        label: String?,
+    ) {
+        val updateApplicationInfoGridItems =
+            applicationInfoGridItemRepository.getApplicationInfoGridItems(
+                serialNumber = serialNumber,
+                packageName = packageName,
+            ).filterNot { applicationInfoGridItem ->
+                applicationInfoGridItem.override
+            }.mapNotNull { applicationInfoGridItem ->
+                if (componentName != null) {
                     UpdateApplicationInfoGridItem(
                         id = applicationInfoGridItem.id,
-                        componentName = launcherAppsActivityInfo.componentName,
-                        label = launcherAppsActivityInfo.label,
-                    ),
-                )
-            } else {
-                deleteApplicationInfoGridItems.add(applicationInfoGridItem)
+                        componentName = componentName,
+                        label = label,
+                    )
+                } else {
+                    null
+                }
             }
-        }
 
         applicationInfoGridItemRepository.updateApplicationInfoGridItems(
             updateApplicationInfoGridItems = updateApplicationInfoGridItems,
         )
-
-        applicationInfoGridItemRepository.deleteApplicationInfoGridItems(
-            applicationInfoGridItems = deleteApplicationInfoGridItems,
-        )
     }
 
-    private suspend fun updateWidgetGridItems(appWidgetManagerAppWidgetProviderInfos: List<AppWidgetManagerAppWidgetProviderInfo>) {
+    private suspend fun updateWidgetGridItemsByPackageName(
+        appWidgetManagerAppWidgetProviderInfos: List<AppWidgetManagerAppWidgetProviderInfo>,
+        serialNumber: Long,
+        packageName: String,
+    ) {
         if (!packageManagerWrapper.hasSystemFeatureAppWidgets) return
 
         val updateWidgetGridItems = mutableListOf<UpdateWidgetGridItem>()
 
         val deleteWidgetGridItems = mutableListOf<WidgetGridItem>()
 
-        val widgetGridItems =
-            widgetGridItemRepository.widgetGridItems.first()
+        val widgetGridItems = widgetGridItemRepository.getWidgetGridItems(packageName = packageName)
+
+        val appWidgetManagerAppWidgetProviderInfos =
+            appWidgetManagerAppWidgetProviderInfos.filter { appWidgetManagerAppWidgetProviderInfo ->
+                appWidgetManagerAppWidgetProviderInfo.packageName == packageName
+            }
 
         widgetGridItems.filterNot { widgetGridItem ->
             widgetGridItem.override
         }.forEach { widgetGridItem ->
-            currentCoroutineContext().ensureActive()
-
             val appWidgetManagerAppWidgetProviderInfo =
                 appWidgetManagerAppWidgetProviderInfos.find { appWidgetManagerAppWidgetProviderInfo ->
-                    currentCoroutineContext().ensureActive()
-
-                    appWidgetManagerAppWidgetProviderInfo.componentName == widgetGridItem.componentName
+                    appWidgetManagerAppWidgetProviderInfo.packageName == widgetGridItem.packageName && appWidgetManagerAppWidgetProviderInfo.componentName == widgetGridItem.componentName && serialNumber == widgetGridItem.serialNumber
                 }
 
             if (appWidgetManagerAppWidgetProviderInfo != null) {
@@ -435,27 +404,33 @@ class SyncDataUseCase @Inject constructor(
         widgetGridItemRepository.deleteWidgetGridItemsByPackageName(widgetGridItems = deleteWidgetGridItems)
     }
 
-    suspend fun updateShortcutInfoGridItems(pinnedLauncherAppsShortcutInfos: List<LauncherAppsShortcutInfo>?) {
+    private suspend fun updateShortcutInfoGridItemsByPackageName(
+        serialNumber: Long,
+        packageName: String,
+    ) {
         if (!launcherAppsWrapper.hasShortcutHostPermission) return
 
         val updateShortcutInfoGridItems = mutableListOf<UpdateShortcutInfoGridItem>()
 
         val deleteShortcutInfoGridItems = mutableListOf<ShortcutInfoGridItem>()
 
-        val shortcutInfoGridItems = shortcutInfoGridItemRepository.shortcutInfoGridItems.first()
+        val shortcutInfoGridItems = shortcutInfoGridItemRepository.getShortcutInfoGridItems(
+            serialNumber = serialNumber,
+            packageName = packageName,
+        )
 
-        if (pinnedLauncherAppsShortcutInfos != null) {
+        val launcherAppsShortcutInfos = launcherAppsWrapper.getPinnedShortcutsByPackageName(
+            serialNumber = serialNumber,
+            packageName = packageName,
+        )
+
+        if (launcherAppsShortcutInfos != null) {
             shortcutInfoGridItems.filterNot { shortcutInfoGridItem ->
                 shortcutInfoGridItem.override
             }.forEach { shortcutInfoGridItem ->
-                currentCoroutineContext().ensureActive()
-
                 val launcherAppsShortcutInfo =
-                    pinnedLauncherAppsShortcutInfos.find { launcherAppsShortcutInfo ->
-                        currentCoroutineContext().ensureActive()
-
-                        launcherAppsShortcutInfo.shortcutId == shortcutInfoGridItem.shortcutId &&
-                            launcherAppsShortcutInfo.serialNumber == shortcutInfoGridItem.serialNumber
+                    launcherAppsShortcutInfos.find { launcherAppsShortcutInfo ->
+                        launcherAppsShortcutInfo.shortcutId == shortcutInfoGridItem.shortcutId && launcherAppsShortcutInfo.serialNumber == shortcutInfoGridItem.serialNumber
                     }
 
                 if (launcherAppsShortcutInfo != null) {
@@ -488,5 +463,78 @@ class SyncDataUseCase @Inject constructor(
 
             shortcutInfoGridItemRepository.deleteShortcutInfoGridItems(shortcutInfoGridItems = deleteShortcutInfoGridItems)
         }
+    }
+
+    private suspend fun updateShortcutConfigGridItemsByPackageName(
+        serialNumber: Long,
+        packageName: String,
+        applicationIcon: String?,
+        applicationLabel: String,
+    ) {
+        val updateShortcutConfigGridItems = mutableListOf<UpdateShortcutConfigGridItem>()
+
+        val deleteShortcutConfigGridItems = mutableListOf<ShortcutConfigGridItem>()
+
+        val shortcutConfigGridItems = shortcutConfigGridItemRepository.getShortcutConfigGridItems(
+            serialNumber = serialNumber,
+            packageName = packageName,
+        )
+
+        val eblanShortcutConfigs = launcherAppsWrapper.getShortcutConfigActivityList(
+            serialNumber = serialNumber,
+            packageName = packageName,
+        )
+
+        shortcutConfigGridItems.filterNot { shortcutConfigGridItem ->
+            shortcutConfigGridItem.override
+        }.forEach { shortcutConfigGridItem ->
+            val launcherAppsActivityInfo = eblanShortcutConfigs.find { launcherAppsShortcutInfo ->
+                launcherAppsShortcutInfo.componentName == shortcutConfigGridItem.componentName && launcherAppsShortcutInfo.serialNumber == shortcutConfigGridItem.serialNumber
+            }
+
+            if (launcherAppsActivityInfo != null) {
+                val activityIcon = launcherAppsActivityInfo.activityIcon?.let { byteArray ->
+                    fileManager.updateAndGetFilePath(
+                        directory = fileManager.getFilesDirectory(FileManager.SHORTCUT_CONFIGS_DIR),
+                        name = launcherAppsActivityInfo.componentName.replace(
+                            "/",
+                            "-",
+                        ),
+                        byteArray = byteArray,
+                    )
+                }
+
+                updateShortcutConfigGridItems.add(
+                    UpdateShortcutConfigGridItem(
+                        id = shortcutConfigGridItem.id,
+                        componentName = launcherAppsActivityInfo.componentName,
+                        activityIcon = activityIcon,
+                        activityLabel = launcherAppsActivityInfo.label,
+                        applicationIcon = applicationIcon,
+                        applicationLabel = applicationLabel,
+                    ),
+                )
+            } else {
+                deleteShortcutConfigGridItems.add(shortcutConfigGridItem)
+            }
+        }
+
+        deleteShortcutConfigGridItems.forEach { shortcutConfigGridItem ->
+            shortcutConfigGridItem.shortcutIntentIcon?.let { shortcutIntentIcon ->
+                val shortcutIntentIconFile = File(shortcutIntentIcon)
+
+                if (shortcutIntentIconFile.exists()) {
+                    shortcutIntentIconFile.delete()
+                }
+            }
+        }
+
+        shortcutConfigGridItemRepository.updateShortcutConfigGridItems(
+            updateShortcutConfigGridItems = updateShortcutConfigGridItems,
+        )
+
+        shortcutConfigGridItemRepository.deleteShortcutConfigGridItems(
+            shortcutConfigGridItems = shortcutConfigGridItems,
+        )
     }
 }
