+++ domain/use-case/src/main/kotlin/com/eblan/launcher/domain/usecase/pin/AddPinWidgetToHomeScreenUseCase.kt
@@ -15,7 +15,7 @@
  *   limitations under the License.
  *
  */
-package com.eblan.launcher.domain.usecase
+package com.eblan.launcher.domain.usecase.pin
 
 import com.eblan.launcher.domain.common.dispatcher.Dispatcher
 import com.eblan.launcher.domain.common.dispatcher.EblanDispatchers
