+++ feature/home/src/main/kotlin/com/eblan/launcher/feature/home/HomeViewModel.kt
@@ -33,23 +33,23 @@ import com.eblan.launcher.domain.model.PinItemRequestType
 import com.eblan.launcher.domain.repository.EblanShortcutInfoRepository
 import com.eblan.launcher.domain.repository.FolderGridCacheRepository
 import com.eblan.launcher.domain.repository.GridCacheRepository
-import com.eblan.launcher.domain.usecase.CachePageItemsUseCase
-import com.eblan.launcher.domain.usecase.DeleteGridItemUseCase
 import com.eblan.launcher.domain.usecase.GetEblanAppWidgetProviderInfosByLabelUseCase
 import com.eblan.launcher.domain.usecase.GetEblanApplicationComponentUseCase
 import com.eblan.launcher.domain.usecase.GetEblanApplicationInfosByLabelUseCase
 import com.eblan.launcher.domain.usecase.GetEblanShortcutConfigByLabelUseCase
 import com.eblan.launcher.domain.usecase.GetFolderDataByIdUseCase
-import com.eblan.launcher.domain.usecase.GetGridItemsCacheUseCase
 import com.eblan.launcher.domain.usecase.GetHomeDataUseCase
-import com.eblan.launcher.domain.usecase.GetPinGridItemUseCase
-import com.eblan.launcher.domain.usecase.MoveFolderGridItemUseCase
-import com.eblan.launcher.domain.usecase.MoveGridItemUseCase
-import com.eblan.launcher.domain.usecase.ResizeGridItemUseCase
-import com.eblan.launcher.domain.usecase.UpdateGridItemsAfterMoveUseCase
-import com.eblan.launcher.domain.usecase.UpdateGridItemsAfterResizeUseCase
-import com.eblan.launcher.domain.usecase.UpdateGridItemsUseCase
-import com.eblan.launcher.domain.usecase.UpdatePageItemsUseCase
+import com.eblan.launcher.domain.usecase.grid.DeleteGridItemUseCase
+import com.eblan.launcher.domain.usecase.grid.GetGridItemsCacheUseCase
+import com.eblan.launcher.domain.usecase.grid.MoveFolderGridItemUseCase
+import com.eblan.launcher.domain.usecase.grid.MoveGridItemUseCase
+import com.eblan.launcher.domain.usecase.grid.ResizeGridItemUseCase
+import com.eblan.launcher.domain.usecase.grid.UpdateGridItemsAfterMoveUseCase
+import com.eblan.launcher.domain.usecase.grid.UpdateGridItemsAfterResizeUseCase
+import com.eblan.launcher.domain.usecase.grid.UpdateGridItemsUseCase
+import com.eblan.launcher.domain.usecase.page.CachePageItemsUseCase
+import com.eblan.launcher.domain.usecase.page.UpdatePageItemsUseCase
+import com.eblan.launcher.domain.usecase.pin.GetPinGridItemUseCase
 import com.eblan.launcher.feature.home.model.EblanApplicationComponentUiState
 import com.eblan.launcher.feature.home.model.EblanShortcutInfoByGroup
 import com.eblan.launcher.feature.home.model.HomeUiState
