+++ service/src/main/kotlin/com/eblan/launcher/service/SyncDataService.kt
@@ -20,7 +20,7 @@ package com.eblan.launcher.service
 import android.app.Service
 import android.content.Intent
 import android.os.IBinder
-import com.eblan.launcher.domain.usecase.SyncDataUseCase
+import com.eblan.launcher.domain.usecase.launcherapps.SyncDataUseCase
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
